author-user-link (#56,#57,#60): more flexible approach to set initial state for automated tests
different features need different type of tests data and tests user,
so the set up of the initial state need be able to configured differently for each feature.
The bare minimul is kept as a before suite hook on the main context,
but the rest will be done in a Background scenario so I wrote the appropriate step definitions
and update the feature files accordingly.