admins attach author to user (#56): implementation done
It is now possible for an admin to attach an author to a user from a user view or edit form.

AdminAuthorController has had its admin action updated to set the user id into an attach_user session property
that determine the display of an explanation message and a "Cancel attaching author" button on the author admin view.
If the value of the session property is "abort" the property is unset and the user is redirect to the normal author admin view.
The author form has been updated to display the user id as default value in the gigadb user field if the sesssion property for attach_user exists.
The view and update form for a gigadb user has been updated to show a "Attach an author to his user" button to initiate the process by sending a attach_user GET parameter
to the AdminAuthorController's admin action.
Created a new AuthorUserContext feature context for step definitions for testing the attachment of author to user by admins workflow.
I particular:
- it make use of Chained step definitions that allow long and complex scenarios to be divided into simpler chained scenarios.
- it has a step for clicking on specific row of a table based on a text