linking author to user (#56, #60)
In order to simplify and speed up Behat automated tests,
we will connect to the postgresql sever in the vagrant box directly
from the php code.

connecting the dev environment to postrgresql inside vagrant box is 3-steps:
setting port forwarding in Vagrantfile is only first step.
Secondly, we need to changes posgresql.conf to get the database server to listen
on all interface, not just loopback interface.
Thirdly we need to add client authorisation for the dev environment in pg_hba.conf
In this commit, I've updated default attributes in Chef to allow just that.
Note that the 10.0.2.2 IP address used in client_auth is a Vagrant "magic" IP address that
represent the host running the Vagrant box.