admins attach author to user (#56,#60): refactoring the acceptance tests
the step to load a user into the database nows connect to the postgresql
 database server in the vagrant box directly from PHP.
Also the sql has been updated into a insert query that can be used in pg_* php functions
instead of a PGSQL command.
a private function actually handles the database loading of a user for cases
when we want to do it without wanting to set the user_login/password.