Add test related containers services and custom postgres configuration
the new test service is non-daemonized, it's meant to be run as
docker-compose run test

It uses the same Dockerfile as for the application service but activates (INSTALL_*)
enough PHP extensions and tools to allow manipulating the database and
restart the application service from within the container.

phantomjs is now automatically instantiated as a container in daemon mode

The postgres container has a custom pg_hba config to allow no-password
access for the test container.

the default URL for the local gigadb web site is now
gigadb.gigasciencejournal.com
to remove dependence on DNSMasq

Updated the production like data dump with the new views from the
optimised new homepage.

Created a pg_restore friendly version of gigadb_testdata.sql for easy
database reset in the test container by the acceptance tests frameworks.

Updated Behat config to use containerized phantomjs and php application