Setup behat and docker to allow test to access webserver oauth callback
From within the same container network, the containers can access each others using service name.

Howeverm for acceptance testing of affiliate login, only redirect urls
registered at affiliate services must be used and they reject the service name
as not a valid url.

The solution is to add in /etc/hosts of the test and phantomjs containers
a valid hostname/ip address combination
(using the extra_hosts option in docker compose, and adding additional server_name in nginx config)

The problem is that IP address is dynamically and randomly created when the containers
are instantiated.

The solution is to use a known fixed IP address for the web server. This is done
by adding an IPAM block to the top-level networks configuration and using the ipv4_address
option in the web server network configuration with an arbitrary chosen IP address
that matches the subnet in top-level network configuration.

So then, the solution becomes to use extra_hosts option to bind a valid domain name
to the known IP address of the web server on the test and phantomjs containers
(the two containers that need to know the affiliate-friendly redirect url)

Also bind-mount ./tmp to /tmp so we can acccess screenshot of failed acceptance tests.