Fix compatibility issues with samples and files tables on dataset pages
Use Recent version of Jquery(2.2.4) and Bootstrap (3.3.7) because needed by the new design.

Stop using CGridView for the tables as that Yii 1.1 widget relies on old jquery that
cause incompatibility with the above combination of JQuery and Bootstrap.

Instead use Datatables library. When we migrate to Yii 2, it will still works.

We still want to use the CActiveDataProvider in the Dataset controller, but
we want to integrate it with Datatables (and with Datatables's Bootstrap theme)
That requires handling the pagination with CPagination outside CActiveDataProvider in the controller.
And it requires handling the pager with SiteLinkPager separately from the tables in the view.

A new CSS file (datatables.css) is introduced to ensure the Datatables tables for Samples and Files
look and feel the same way as before in the original version of the tables.

Two inline Javascript scripts (one for samples, one for files)
are necessary to enable Datatables on the HTML5 tables rendered from CActiveDataProvider
as well as for disabling Datatables features that weren't present in the original version of the tables.

Remove redundant special case for dataset model 629, as it was there
likely for debugging purpose and then likely forgotten.