Enable HTTPS on local development machine for FileUploadWizard
Using minica set up as a container, create certificates pair and root certificate
to be used by nginx container.
Modified Docker compose file's local override to make the above happen.

It's required that the root certificate be added to the developers' web browsers
for this to work. It also require to use a domain that hasn't got HSTS enabled.