Fix new DB schema related issue and streamline legacy acceptance runner
before restoring the original database after running acceptance tests,
we need to drop the database because some of test database dump loaded
during test run have out-dated schema that cause the post-tests
pg_restore process to fail. Due to the need for  all clients to be
disconnected before dropping database, the drop and restore has to be done
as a AfterSuite hooks in Behat when we still got a chance
to kill connections and restart php container. It's also more elegant
than doing it in the bash script.