Reduce significantly time to build custom nginx Docker image
it was very slow because it was configured to use the root of the project as context,
which required the whole git repo to be loaded in context memory.

Moved public assets for GigaDB out of root dir into ops/configuration and
select ops/ as context.  Build time: 41mn -> 5mn

TODO for future: Create a proper sub-dir for GigaDB webapp's application and assets files