Acceptance test setup and first scenario's steps implemented and ready
The test scenario is failing for the right reason (business feature not yet implemented)
we are now ready to implement it.

codeception is now configured so we can load test data (and it cleans after itself) in the existing database
in the Background steps' implementation.

For now we assume the database is already setup for running acceptance tests.