Add a token service class to generate JWT token for API calls (wip)
It will be used (and is therefore a pre-requisite) by the FiledropService
that powers the UX for assigning file upload dropbox to a dataset.

It is part of the GigaDB's Yii1-based codebase and test harness but uses Yii2 component system
(inheritance and configuration).

It's still work in progress. UserDAO is not implemented yet.