Add a Messaging Service to File Upload Wizard
Goal: This service's sendEmailMessage function will be called
by filedrop-accounts controller's update action when accompanied with email related query strings.

The service is implemented as a Yii2 Component in the backend (admin) webapp,
it uses yii2-swiftmailer included in the Yii2 framework for sending mail.
yii2-swiftmailer is configured in Yii2 config files (main-local.php.dist, test-local.php.dist).

In the future the service would be place to handle other kind of messaging.