Configure containers for FUW uploader, add file upload button to profile
Check out the @ok acceptance scenarios in the commited feature file to see what's been implemented.

A new php-fpm container service is started in docker-compose file to run File Upload Wizard's frontend code
In yii2 advanced template layout, that's the application aimed at running public accessible code for the authors
(as opposed to backend code which in our case is the filedrop account management api and accessible by admin only)

The new webapp is made available in GigaDB under /uploader/ through site configuration for Nginx.

Also fix the Bootstrap tabs on the user's profile so the the right tab is loaded
based on the hash string in the url.

still work in progress and some dummy code to remove.