Fix issue with keywords validation in Dataset attributes
This looks like and old bug rather than a regression. It may happen on production too.

It is connected to the bug of the redirect not working with long urls,
as in the test for this didn't fail because of the same root cause of the redirect issue (see previous commit), 
but this time the dodgy characters entered in the test scenario was longer than the artificially limited field length
so the dodgy keyword was indeed rejected, but for the wrong reason.

Added proper validation and unit tests. The acceptance scenario for this feature doesn't need fixing.