Remove defer attributes from JS includes on a legacy page to avoid error
the user admin.php page use the yii 1 style zii widgets to display the table
that use javascript and jquery that need to be loaded when the code is parsed.

To use defer, I'd have to refactor the whole page and the test mink context
which was not the current intent.