Implement file upload function in GigaDB using File Upload Wizard
It was necessary to update the gigadb.dev.http.conf to allow the upload requests
to be proxied through to the upload server Tusd.
it is particular important to setup up proxy_set_header to the Nginx special variables
$host:$server_port as the javscript uploader need to query the upload server regularly
to track upload progress and those values differ based on the environments and whether they are accessed from test suite or for real (so cannot hard-code or configure them easily).

Because the acceptance tests need to use File Upload wizard's file drop account for real,
we need to be able to create and delete the directories used by such account.
The directory creation is in the implementation of the background step "@Given filedrop account for DOI :doi does exist" in AuthorSteps.php, using the Codeception's Filesystem module.
But for the cleanup we need to implement an after test hook.
This is what the new file FilesystemCleanupHook.php is for.