Fix GigaDB http config for file upload
with previous settings tests are passing
(they use gigadb.test domain and the file chunks's location are set to gigadb.test:80/files/... by tusd upload server)
but actual file upload on dev failed because the file chunk's location are set to
gigadb.gigasciencejournal.com:80/files/...
which the web browser doesn't know about.
using the nginx special variable $http_port allow setting of the chunk's location to whatever
was used in the originating request:

gigadb.test:80 when called from acceptance tests

and gigadb.gigasciencejournal.com:9170 when call by a human in dev environment.

(the problem was the use of $server_port which refers to the actual port
used in the Listen directive of the nginx config file, which is 80)