Implement access control to file uploading endpoints in GigaDB
Ensure, at the controller level, that only the submitter of the dataset can upload files
and only when the dataset have the correct upload status (UserUploadingData).
Other logged in users are forbidden (403 http status code)
and if wrong status (409 http status code) is returned

Not logged in users who load the file uploading urls will be redirected to the login screen.

The above functionality is implemented using Yii 1.1's controller filters.

See:
* https://www.yiiframework.com/doc/guide/1.1/en/basics.controller#filter
* https://www.yiiframework.com/doc/guide/1.1/en/topics.auth

Automated functional tests accompany the implementation.

DatatabaseSteps.php is a trait with helpers methods for setting up database state before test run.

The functional test suite for GigaDB are susceptible to memory leaks (another reason to switch to Codeception)
so I made sure the tests are not source of memory leakage, so I paid attention to
the setUp and tearDown methods of the test class, making sure everything that has been opened in the test
are close properly and parent setup and tear down are called at the right position.