In functional tests, fix wrong configured database and a race condition
RSSFeedTest.php was occasionally randomly failing because the order of the feed item could vary
base on running context, because of the sorting criteria causing two items to be on same rank
and yet the test assertion made assumption about which one comes first.

Fix DB config for functional test to use the main database. It was erroneously configured
to use test database which is for unit testing.