Add API endpoint to File Upload Wizard REST API to create a new user
This is so to allow the author using for the file uploader and metadata form.

Because the User class is the User identity used for authentication and by default a new
user filling a sign up form that need email-based verificaiton, I needed to create the
Yii2 scenario (admin-created) that allow the creation in situ of brand new User object
that's also active by default.