Update Filedrop creation to add author in FUW DB on sending instructions
As part of uploading files and annotating them with meta-data, the author
will need to interact with File Upload Wizard REST API, so the author's login on GigaDB will need to be authorised
with FUW and therefore their credential should be in in FUW database's table for authorisation (public.user).

The appropriate time for creating such record, is at the end of the filedrop account
creation process by the curator, when the email instructions are sent to the author.

The important part is in the emailInstructions() function of FiledropService and it has required the changes to DatasetDAO and TokenService
from previous commits (see 9504afed and b364af39)

Because the return value of the above function has changed to be more useful, I had to change DatasetUpload
(the badly named DTO class that link GigaDB and FUW for filedrop account management, got to change that name later - TODO - )
to work with the function's new signature.