Extend Codeception\Db to remove Db record added by FUW code in BDD tests
The recent change introduced as a prerequisite to the metadata file annotation form work,
to create a user in File Upload Wizard database for the dataset author
to authenticate and authorise them accessing FUW REST API has broken the acceptance tests.
Because a new DB record is added by the production code that is not deleted after test run
the test breaks when run the second time.

The acceptance tests use Codeception's Db module for handling DB interaction is able to
delete the record it creates (e.g: haveInDatabase()) but not the records created by production code.

So I create an module that extend Db to remove that record created by the production code
and in the acceptance tests configuration file, I drop-replaced Db by the new extension.

The code used in the extension takes inspiration and follow the coding style seen in the parent Db.php module.