Dockerfile for infrastructure to test and build Javascript components
This will allow running unit tests for the Javascript components used in
GigaDB/File Upload wizard integration in the same manner as PHP unit tests.

It also remove the needs for developer to have node/npm installed on their
development machine, which in turn allow us to not have to think of
platform idiosyncrasies between OSes (macOS, Linux, Windows).

Note that this Dockerfile only sets up the infrastructure:
the latest version of NodeJS ecosystem and a headless web browser
on the current Debian distribution.
It does not make any assumptions about the application framework
or test framework, nor does it care about deployment and assets pipeline.

However it expects to find the Javascript codebase under 
the /var/www/gigadb/app/client/web directory structure, as that's where the WORKDIR 
command is set to.