Add docker-compose service to test, build, deploy Javascript components
Using the new custom Dockerfile for NodeJS ecosystem and headless web browser,
the new service's default command will run the test, build and deploy scripts
specified in the package.json project configuration file of the Javascript components codebase.

The command will be run from the WORKDIR value set in the Dockerfile
which is /var/www/gigadb/app/client/web, so we need to bind-mount the
application directory to /var/www, so that the Javascript codebase can be found.

Containers that run desktop web browsers need to have the shared memory setting (shm_size)
configured otherwise they'll throw errors.

The Container image run Debian Buster, so it's possible to use bash for running commands
and for opening a debugging terminal