Fix random acceptance tests issues by replacing phantomjs with chrome
phantomjs is increasingly flaky, so we switch to another headless browser.
Unfortunately SeleniumWebDriver support for Firefox has
non-workaroundable or no-upcoming-fixes-soon problems [1] ,
only Chrome up to version 74 works [2], but it's a better place to be.

By switching to Chrome, we have to alter (actually improve) a couple
of background scenarios.

Also *.dev domains are restricted in Chrome [3] so switching the internal
domain to *.test.

[1] https://stackoverflow.com/questions/46913377/php-codeception-not-so-quick-start-phpunit-framework-exception-undefined-inde
[2] https://github.com/Behat/MinkExtension/issues/345
[3] https://www.theregister.co.uk/2017/11/29/google_dev_network/