Add first passing unit test for the file uploader JS Vue component
The file uploader library (Uppy) we use is only available as a React component at the moment,
since file upload wizard frontend is built on Vue.js, we need to wrap Uppy within a Vue.js component.

With this commit, we have a basic Vue component for the Uppy uploader and we've tested successfully that
the drag-and-drop area is loaded in the rendered component.