Add required configuration parameters to the Vue.js UploaderComponent
The identifier (DOI) and the TUS endpoint are information mandatory for
the proper functioning of the Uppy uploader in GigaDB, the later so that
the uploader know which remote endpoint to send the files to, and the former
so that we know with which dataset an uploaded file is associated with.

Their value come from the server side code (PHP), so they need to be
passed to the component as parameter (props), as the component will be used
on the view file for the authorisdedDatasetController's fileUpload action.