Fix inconsistencies and brittleness in the JS unit tests
Because we are testing the behaviour of the Uppy component after its host
mounted and connected to the DOM, we need to ensure our test wrappper for
the host component  has "attachToDocument" set to true,
and all our expectations need to be wrapped in Vue.nextTick() to ensure the
test wrapper has completely loaded before we probe the rendered structure.
Also removed duplication in the test wrapper factory.