Replace direct use of Uppy by wrapping it in a Vue component
This is so because we cannot have the page being a proper Vue application's in-document template,
and separately instantiate the Uppy library in its own <script> tag.
Additionally we need to handle in our Vue application events emitted by the Uppy library.
So, instead we've wrapped the Uppy library in the UploaderComponent Vue component with events pass-through
and use the associated custom element in the authorisedDatasetController.php filesUpload action's view page
(filesUpload.php).