Use event bus for Vue components to communicate and improve unit tests
Because the uploader component needs to inform the future pager sibling component
that files upload have completed, we need mechanism for siblings to pass on event to each others.
(the default event mechanism is for child to parent relationships)
Using an event bus is a common pattern thats suitable to use here.

Because it needs to be tested, the event bus will be passed to components that
needs to communicate using Props.

Also I figured out how to more accurately tests that we are bubbling out the events
generated by uppy. Now we can force uppy to emit an event in our unit tests.
It also mean we can get rid of the intermediary method in the uploader component