Add endpoint to File Upload Wizard API to return file uploads for a DOI
First, delete all the unnecessary files for the dummy application installed by Yii2
when creating an advanced template project.

The endpoint is implemented as UploadController for the Upload model
which is almost text book example for ActiveController, Yii2's REST endpoint library.

Useful resources:
* https://www.yiiframework.com/doc/guide/2.0/en/rest-resources
* https://www.yiiframework.com/doc/guide/2.0/en/output-data-providers#filtering-data-providers-using-data-filters
* https://stackoverflow.com/questions/58755547/yii2-yii-data-activedatafilter-and-pagination-false

The latter is where the main controller code is taking inspiration from.

Also added a missing column to the upload table for the data type

And configured Nginx web server to proxy REST requests to the new endpoint
running in the fuw-public container
It has to be separate from fuw-admin container reserved for admin privilege endpoints, for security (and
 matching Yii2 advanced template project structure).