Fix issues preventing PHP to pass on list of uploads to Js Vue component
Firstly, there's a delay between the file is uploaded in the browser and the file meta-data
being generated and saved in the database, so we add a 60 seconds wait in the acceptance test scenario.

Secondly, we had a typo in the filesAnnotate.php view template and missed the semi-colon.

Thirdly, the list is a PHP array formated as JSON array. They use double-quotes as delimiter
around keys and values, so the double-quotes surrounding the PHP snippet in the value of  ``uploads`` attribute of the 
annotator component in the view template
caused the array to obliterate the rendered component, we fix it by using single-quote.