Add form elements to the metadata form
For each of the uploaded files ( a list returned by the PHP controller)
we build the form element that show the current default as well as the appropriate
form element when it makes sense: Data type (as the default was only approximately guessed)
menu with the same options as in file-to-db.php (ideally there should be only one place where they are defined, todo for future refactoring)
and the description's free text input.

each form element as an id indicating its index in the list of uploaded file and the type of data, notice that the index start at one in the
acceptance scenario, but at 0 in the AnnotatorComponent.vue's template.
This is because in the acceptance tests scenario, the table has a heading wich takes index 0 already,
so all the list of files to test against  have index starting at 1.
in the Vue component, the list of uploads is just the list without heading, thus starting at 0.
However in template we want to use natural index starting at 1 as it make more sense, so we do (index+1).

Bootstrap is used for the styling of the form. Because it's used all over on GigaDB, there's no need to import it into the Vue root applicationm (index.js).

Although the acceptance tests scenarios realated to the meta-data form are still under the Author role, the AuthorSteps.php scenarios implementations
is already quite big, so I'm creating a new MetadataFormSteps.php specific for corresponding scenarios.