Allow Annotator component to notify when all uploads' fields are ready
Knowing when the user has filled info for all uploads, by emitting an event
will allow controlling the display or not of the button to move to the next stage

This done by capturing events from changes to the form fields and keeping a tally of completed uploads.

in the test, when instantiating a component, make sure we pass a copy of the list of uploads
(using ``JSON.parse(JSON.stringify( uploads ))``) otherwise the original list will be
modified and break isolation between each tests, causing unexpected values in the list.