Fix tests for the next button to move from uploading to annotating stage
starting from the last couple previous commit, we've got a really better understanding on how
to test event-driven, reactive component changes with Vue and Jasmine.

In addition, we've replaced the use of variable local to the ``describe`` block
for holding our component's test wrapper by using Jasmine context (``this``)
as it allows us to benefit from Jasmine variable auto-purging after each call to the ``it`` blocks,
thus preventing memory leaks.

Reminder from last commit about why we change the content of the ``it`` blocks:

>previous commit had annotator component to emit the metadata-ready-status event
>when all uploaded files have their metadata ready, here:
* we update the Pager component to subscribe to that event,
* we update the Annotator component to notify the current stage (annotating)
* we update the Pager component's template to display the completion button
   when stage is annotating and metadata-ready-status is true

>Regarding the unit tests, it uses three tricks in order to function properly:
* Because the template is not re-rendered immediately when component data is updated, but after the next ``tick``, we need to wrap
  our test expectation inside a $nextTick closure [1]
* the ``this`` object cannot be used to refer to the caller's context inside a closure
   therefore we need to assign the component wrapper associated to ``this`` context to a new variable
   that will be used in its place inside the closure.
   This is so because in closures, ``this`` is associated with the local context
* We use Jasmine as our test framework, it has good support for asynchronous code through Promises.
so our closure needs to use the Promise version (instead of the callback version), and we need to ``return`` the closure [2]

[1] https://vuejs.org/v2/guide/reactivity.html#Async-Update-Queue
[2] https://jasmine.github.io/tutorials/your_first_suite#section-Using_promises