Complete saving of updated metadata for all uploaded files
This is the implementation of thew new acceptance scenario for saving metadata.

On the client side, it involves submitting the metadata form made of Vue components:
AnnotatorComponent.vue for the tabular data form elements and PagerComponent.vue for the submit button
which means the <form> tag needs to be in the filesAnnotate.php view so it can wrap both components.

We then needs to update the FilesAnnotateAction.php to process the form submission.
What we do there with the data is to send it to the File Upload Wizard API's REST endpoint
for  updating uploads, and we do so for each uploaded files.

To achieve that, we need to add a UpdateAction.php to the FUW app (not strictly necessary
as there's currently no custom behaviour for update uploads but useful for debugging)
and update the UploadController.php to allow ``update`` action and register UpdateAction.php

We also need to update FiledropService.php (GigaDB's client code to interact with FUW API)
to add the function ``updateUpload`` that will allow FilesAnnotateAction.php to send the form data to the API.