Add Dataset status change, notification when uploads metadata's complete
The FilesAnnotateAction.php in charge of handling metadata form for File uploads
now sets the Dataset upload status to "DataAvailableForReview" and send and email to
the editorial email address (configured in params-local.php.dist).

The code enabling the above is implemented in the new business domain layer class DatasetUpload.php.