Add UI for collapsible panel to add attributes to file uploads' metadata
use a drawer (collapsible panel) [1] as foundation for the form (yet to implement) to add new attributes
to the metadata form for uploaded files

This include triggering the drawer when the user clicks the "Attributes" button in the AnnotatorComponent.vue and
adding the code for the drawer in the same component.

It has also required to update the Javascript build process so that the assets used by ElementUI UX framework,
are copied in GigaDB's ``assets`` directory so they can be addressed in the rendered and deployed User Interface (UI)
of the metadata form.

The actual form to add attributes that we'll lay on the drawer will be managed in the SpecifierComponent.vue
(named as such because that's where the author specifies the attributes associated to the uploaded files)
for which we've created a skeleton code (it's still work in progress).

To enable <specifier> to be used in the Annotator component, we need to import the new component
and register it within the AnnotatorComponent.vue's <script> block.

[1] https://element.eleme.io/#/en-US/component/drawer