Add functional test for GigaDB's FilesAnnotateAction
This is appropriate and useful level for test-driving the submission of both upload metadata and
associated file attributes.

The test controller action didn't have one already for testing the implemented browser/controller interaction for
posting just the metadata associated to uploaded files, so we need
to lay down that one first before testing the posting of metadata+file attributes.

That work highlighted a couple of issues in the test AdminDatasetSendInstructionsActionTest.php
that required fixing:
* It wasn't explicit in the test that the tested action was creating authorized identity for author after
filedrop account was created and an email sent to the author
* the test wasn't properly cleaning up after itself (in ``tearDown``) regarding user identity created in FUW database
(as a result of the first point, most probably)