Add skeleton code for /attributes endpoint in File Upload Wizard API
The endpoint will be used by the metadata form to store the attributes set
in its collapsible attribute panel.

We don't use standard actions from Yii2 REST (create, update) because the approach we take
is to delete existing attributes for a given upload and post the new set of attributes from the form.
(allow a flexible and user-friendly UX for the form).

Therefore we are implementing a custom action, ``replace``, that takes the id of and upload object as parameter.