Implement attribute replacement for uploaded files in File Upload Wizard
Because in the UX the author will be able to delete/add attributes at will and
the number of attributes associated to an upload is reasonable,
it's more efficient on the server-side to delete existing set and add the sets
arranged in the Javascript interface.

The ReplaceAction.php is the class where the implementation resides.

Although relatively simple logic, the challenge is that Yii2's ``Model`` form data processing functions
expect the data send through POST request in the body of the request to have a specific structure that I need to figure out
otherwise the  ``Model::loadMultiple`` and ``Model::load`` (the latter being used by the former) won't work properly

An array representation of the form data can be seen in the functional test (FileUploadServiceTest.php)
for the GigaDB client to the FUW public API (FileUploadService.php).

The raw POST representation of the form data can be seen in the functional test (AuthorisedDatasetFilesAnnotateAction.php)
for GigaDB's controller action for the metadata form (FilesAnnotateAction.php)

On File Upload Wizard's side, we implement an endpoint in the public API for attributes (AttributeController.php)
 alongside its action to replace them (ReplaceAction.php), and the corresponding test suite (AttributeCest.php).