The metadata form's attribute panel can display existing attributes
Update acceptance tests scenarios for uploads file attributes management (file-upload-4-author-meta-data-tagging.feature)
to test that already stored file attributes for uploaded files are displayed in the
attribute form. That required adding one new background step to ensure test data is loaded for attributes
(implemented in MetadataFormSteps.php)

Because I wanted to generate the test data attributes' field with random string using Yii2's builtin random generator,
I needed to access the Yii2 webapp singleton (Yii2::$app) which is normally only loaded in the web app, not in the acceptance tests.
To enable Yii2::$app in acceptance tests, I add the Codeception's Yii2 module to the ``acceptance.suite.yml`` configuration file
for acceptance tests.

The Specifier Vue.js component implementing the attributes form already support it (from past commits)
and a endpoint and GigaDB client method to get attributes from the File Upload Wizard's REST API
was implemented in the previous commit. What was left to do was connecting the dots
to integrate all actors in the workflow:
* the GigaDB controller action that get data (uploaded files and now the associated attributes)
into the metadata form (FilesAnnotateAction.php)
* filesAnnotate.php view file to the GigaDB controller needs to pass on the attributes from PHP array to JSON into the Annotator Vue.js component
* The AnnotatorComponent.vue which needs to pass on the server-side attributes to the chain of child components
(ElementUI's ``el-drawer`` which needs to pass it on to the Specifier Vue.js component)

Because we need to test that the button component (ElementUI's ``el-button``) in AnnotatorComonent.vue
selects and passes on the index and upload id of the clicked-on file upload,
we need to switch from Vue unit tests tools ``shallowMount`` to ``mount``
because shallowMount doesn't allow setting expectation on sub-components (because it's "shallow" for performance)