Add delete button and logic for removing uploads from metadata form
In the AnnotatatorComponent.vue, use ComponentUI's icon button to display on file table in the metadata form
as well as the Vue.js method that will remove the clicked-on upload form the uploadedFiles array compnent data and add it
to the filesToDelete array component data.

Also for each entry in the filesToDelete array component data, add a hidden text input, so that submitting the form will
allow processing the delete list on server-side alongside uploads and attributes.