Add server-side support for deleting upload files from metadata form
the uploads are not really deleted. Instead they are archived. This allow
us to not have to deal with resource expensive/complicated filesystem operations as part of the file metadata annotation
workflow, but from an user point of view, it opens the possbility of recovering from user error if there's a need to resurect
a deleted file. The generated filedrop account housekeeping will take care of the archive files.

the uploads have a status. status of 2 means the file is archived. In this stage/workflow, active uploads have status of 0.

I had to modify the ``getUploads`` to not return archived files,
as well as the tests setups that required uploads as test data to have status of 0.

Also for the FileUploadServiceTest.php, replace parent class FunctionalTesting (that use mink/phpunit integration) by PHPUnit's CTestCase,
and remove some unused traits as
the tests doesn't require the features form mink, behat-mink, and the mink/phpunit integration
is very buggy in most recent versions of PHP.