Update metadata-form GigaDB controller and view to process deletions
this is the integration of the few previous commit.
The GigaDB controller for the metadata form can now process the list of deletion
by receiving the list from the Js component and passing it to the FileUploadService.php's method
that interact with FUW API.

The  Gigadb user profile view is updated to display a confirmation message.

And I've added the acceptance scenarios testing the upload removal feature.