Add method to restart container in Docker PHP API
The primary reason is to fix a race condition between DB inserts of
test data and querying them from PHP app in the acceptance tests scenarios for
files attributes form of the File upload wizard.

But eventually, this will (eventually) allow replacing of the script ``restart_php.sh``
used for the same reason in the GigaDB's acceptance tests (GigadbWebsiteContext::containerRestart)

apart from the simple code change (thanks to Docker-PHP library for having a method for restarting containers),
we need to ensure the configuration parameters for DocklerManager are available in the ``common`` module of the File Upload Wizard's project structure.