Fix issues with upload test data for acceptance test scenarios
First issue is wrong status to the file uploads DB records created as fixture data
when acceptance testing files attributes for uploads.

Second issue is a occasional race condition will cause acceptance tests scenarios
to fail, the ones that use uploads db inserts in precondition because the DB inserts
may not be fully commited by the time when the next steps trigger the php app
running on fuw-public container service make a request for the /uploads end point.
The solution are to restart the fuw-public container after the insert
(that's the main reason why I've extended DockerManager to add a container restart function in previous commit),
and move the precondition steps form the background block to each acceptance scenarios.

Third issue is removing the slow steps for uploading two files into the UX of GigaDB file upload wizard
for the acceptance scenarios where it's feasible to use preloaded test data.