Fix race condition in file upload process in File Upload Wizard
To avoid unnecessary repeat processing of uploaded files by
the Watcher container service that continuously watches changes to the filesystem,
we have a system of flag per dropbox that's written once the uploaded files
are added to the database. However that process is controlled by a test
comparing filesystem states which may present race conditions
(revealed by random failure of related acceptance tests)

To avoid that, we write the flag, only after the having ensured that the files
have actually been updated in the database.