Fix test data for File Upload Wizard acceptance test scenarios
TWo issues,
Firstly, test data were not cleaned up properly in case of test failure
or interruption of the tests execution.

Fix to that are the changes in DbExtendedWithHooks.php

Secondly, inserting test data in the database using Codeception DB module's ``haveInDatabase``
method sometimes doesn't work: when accessed from the scenario steps, the data appears to randomly be not available.
Not fully certain whether some rows are not inserted in the first place, or the post-scenario auto-deletion of records
inserted with ``haveInDatabase`` quick in at the wrong time. The assertions I put just after adding each row seems
to indicates it's the latter?

the do-while loop around the DB insert improved reliability a bit, but they're likely be more fixes coming.