Add support for TSV format in bulk metadata uploading from spreadsheet
The most standard way to recognize and map file format and file extensions is the IANA MIME types standard, so we configured
in GigaDB config file for the Yii2 framework the list of mime types supported by File Upload Wizard bulk upload
(the mime types for CSV and TSV files).
Then we modified the business object (DatasetUpload.php) where spreadsheet parsing is done for metadata
to detect the file mime type and to generate an error if the format is not one of the above.

Secondly we need to tell the spreadsheet parsing library to expects a TSV by setting tabs as delimiter
if the input file detected is of the TSV mime type.

In the unit test, we change the way temporary file for test data is created, so we can add a file extension,
otherwise file format detection cannot happen. (We use yii2-flysystem library to help with that).

Then we add an acceptance test scenario to cover the support for TSV files.