Validate against GigaDB file formats in Upload model and upon file upload
Upload.php is the Yii2 ActiveRecord model that handle (de)serialization of Upload data in
the database, the file format (extension) field's value needs to be validated to be from the reference data list
of GigaDb file formats.

The upload database record is actually created asfter the file has been uplooaded, so we need to ensure use
a recognized GigaDB file formats when the function creating the database record tries to guess the file format.

That requires loading the json based reference data in to the file upload hook (file-to-db.php) and before that, and therefore
the watcher container service needs to be configured to access the volume path where the json files are.
Also we need to ensure watcher container service which is built upon Alpine Linux gets the appropriate library to do Json with PHP (php7-json).

The functional test for Upload controller (UploadCest.php) also needs to have its test data updated to use GigaDb file formats (upload_data.php).

Also fixed an issue (in UpdateMultipleAction.php) where all uploads, including archived ones where loaded for merging with the spreadsheet meta-data.