Add curationlog entry after dataset status set to DataAvailableForReview
Changing the upload status to to DataAvailableForReview was already sending an email,
but we forgot to create an entry in the curation log table. There was an acceptance tests scenario
to remind me of that.

The acceptance scenario must include steps associated with author because the status change to DataAvailableForReview
is performed by the author, but the business feature is for the benefit of the curators.

Therefore, we need to alllow the curator role in ``acceptance.suite.yml`` to access the steps implementation for Author steps and metadata steps.
We need to ensure that the CuratorSteps is last in the list as there are some identically named steps and the last implementation wins.

The change in the view template ``admin.php`` is so that we can find the xpath leading to that button in the step implementation in ``CuratorSteps.php``

The change in ``DatasetDAO.php`` is so we can retrieve dataset id that's required by the existing Curation log  method, as the ``DatasetUpload.php``
business object only operates with the identifier (DOI).