Redo the creation of mockup url to be time-limited and tied to reviewer
We use JWT tokens (same tech used for authenticating File Upload Wizard REST API) to embed reviewer and enforce
time limitedness, hence a new method in ``TokenService.php`` for generating the JWT token out of reviewer email and
number of months of validity.

The token is too long to be used as part of a url, so we generate a random UUID (following RFC 4122) to use in URL,
which means the File Upload Wizard needs to keep a mapping between JWT token and UIID,
which is what ``MockupUrl.php`` model class (and the corresponding database table) is for.
We use a third party library (Ramsey\Uuid) to handle UUIDs.

We modify the dataset editing form ``adminDataset/_form.php`` so that the button for generating a mockup
triggers a dialog asking for reviewer email and number of months of validity of the url.

So the corresponding ``MockupAction.php`` controller action is updated to process those two parameters
and generate a JWT token using a new method ``makeMockupUrl()`` on the FUW API client class ``FiledropService.php``.

That method also needs to be passed a new instance of TokenService for token generation
(cannot use the same one passed to FiledropService's constructor used for JWT authentication token).

On FUW side, we create the REST endpoint MockupUrlController.php to give A REST api to the ``MockupUrl.php`` with setup in ``backend/config/main.php``, this is pretty much boiler plate, there's nothing specific to it except for Bearer authentication using JWT tokens like any FUW endpoints.

Because of the new dialog box in the form, I need to implement a new test scenario step in CuratorSteps.php