Add initial impl. of new action on DatasetController to show mockup page
In order to display the dataset mockup page, the reviewer browsing to the URL generated specially for him,
we add a new public action (``MockupAction.php``) to  ``DatasetController.php``.
The mockup will reuse the same flow and most of the same components as the regular view action of the same controller
except that the mockup only shows up for dataset set as Submitted (``DatasetPageSettings.php``'s ``getPageType()`` returning "mockup")
and if the randomly generated UUID (url_fragment) is mapped to a JWT token (containing reviewer's email, dataset DOI and max validity for the url)
in File Upload Wizard API.

The user story for the acceptance tests scenarios for this feature are from the point of view of the reviewer, so
we have added a new role (reviewer) in the acceptance tests configuration file (``acceptance.suite.yml``) and created a new steps implementation class
``ReviewerSteps.php`` for implementing the steps specific to that feature file (``file-upload-6-reviewer-access-drop-box-files.feature``)

Updated File Upload Wizard's endpoint (``MockupUrlController.php``) for getting mockupUrl to throw an exception ("404 Not Found" HTTP exception) if it can't find a record
for the given UUID (url_fragment).