Update user creation and token creation for mockup page
The same reviewer may review more than one dataset, but only need one entry
in FUW user authentication database, so we modify the function that create user in by GigaDB Admin
(``createUser()`` in ``TokenService.php``) to first lookup in the FUW API if the user exists already
(that's what the ``LookupAction.php`` is for) and return the existing user, otherwise it creates the user
by sending a POST to the ``/users/`` endpoint.

We also modify the ``generateTokenForUser()`` in ``TokenService.php``, so that it doesn't necessarily require a
GigaDB user instance, when we create an authentication token for the reviewer as needed in the mockup page to query
the uploaded files from FUW API, this is because reviewer don't necessarily have a GigaDB account.
Accordingly, the ``getUploads()`` method in ``FileUploadService.php`` is updated to use ``generateTokenForUser()`` with
the new signature, and then we instantiate the FileUploadService instance with the newly required property in ``MockupViewAction.php``.