Add changes allowing dataset mockup page to show basic dataset info
enabled the code branching between database and API in ``DatasetPageAssembly.php``
to select API (to retrieve the uploaded files) in mockup view and the files
from the database in regular view.

Changed the steps implementation in ``AcceptanceTester.php`` to have the proper records in database
to allow ``StoredDatasetMainSection.php`` called in ``DatasetPageAssembly.php`` to return the correct dataset info.

Fixed ``ResourcedDatasetFiles.php`` to implement the DatasetFilesInterface interface so it's
interchangeable with ``StoredDatasetFiles.php`` as far as ``CachedDatasetFiles.php`` is concerned
(see ``setDatasetFiles()`` in ``DatasetPageAssembly.php``)

The corresponding acceptance test scenario is implemented too, see:
* file-upload-6-reviewer-access-drop-box-files.feature
* ReviewerSteps.php
* AcceptanceTester.php