Add distinct file settings for mockup and regular view for datasets
Although I want to show files attributes columns when showing mockup view,
we don't want to alter the current default settings (not showing attributes)
on the regular dataset view, even though the code is shared across both.

We also don't want the ``DatasetController.php`` to know about the details about columns names.
That should stay encapsulated within ``DatasetPageSetings.php``.

So we create two constants in that class and add a parameter (with default value set to regular view's columns) to ``getFileSettings()``,
so that we can select the appropriate set of columns within each action
 ``mockupViewAction.php`` or DatasetController.php's ``actionView()``