Support curators data validation of uploaded files
When the author has finished uploadings files and the after the editors set the dataset status to "Submitted"
curators can sanity-check the files and set the status to "DataPending" if some files have problem
(an email is automatically sent to the author) or to "Curation" if all is good.

For acceptance test scenario, there was no need to implement new steps as the steps used
in ``file-upload-6.5-curator-sanity-check.feature``  are the same as in ``file-upload-4.5-curator-editorial-acceptance.feature``

Also refactored ``CurationLog.php`` to remove duplicated code.