Implement the server-side of MD5 checksum of uploaded file
The client/server file uploader component (uppy/tusd) writes the client-side generated MD5 checksum
in the metadata JSON file (``<tusd Id>.info.json``) associated to each file being uploaded.

So the first thing we do is to copy that JSON metadata file alongside the files into a dataset-keyed directory,
hence the change to the ``post-finish`` TUSd hook. The JSON file goes into a "meta" sub-directory of the
dataset-keyed directory so we don't have to change the way the dataset files are being detected by the "watcher"
container service. And the JSON file is renamed to bear include the name of the dataset file is associated with.

Then we update the "fileMetadata()" function in the watcher's script ``file-to-db.php`` to parse the JSON file for the checksum
and add it to the metata array made for each uploaded file.

The we update the "updateFileTable()" function to store the checksum in the upload database table's "initial_md5" column.

Finally, we update the Metadata form in the Vue.js application ``AnnotatorComponent.vue`` to show the checksum as a tooltip when
hovering over the file name.

We have an acceptance scenario for this feature in ``file-upload-3-author-meta-data-form.feature`` which required a new step
to be implemented in ``AuthorSteps.php`` to detect the checksum tooltip.

The ``UploadSearch.php` Yii2 Search model for uploads was updated to allow the initial MD5 Checksum to be use as filter.
That will be useful later when searching for files associated to a particular checksum.