Add button to meta-data form component to add sample ids from a panel
use same approach as for adding file attributes:
In the meta-data form (implemented in ``AnnotatorComponent.vue``), a button next to an uploaded file table's row opens up a panel (<el-drawer>, a Vue component from open-source components library Element-UI) with a form for adding the samples.

The form comes from the previously implemented SamplerComponent.vue (see previous commit).

Some problems I had to solve here:

* In order to save the entered sample ids into a uploaded file's sample_ids attribute, we have the ``SamplerComponent.vue`` to emit
an event upon a click on the "Save" button, then ``AnnotatorComment.vue`` which has a subscription on that event will concatenate
all sample ids into a string suitable to be saved in the database's upload record's sample_ids attribute.
The event needs to be emitted by the ``SamplerComponent.vue`` and not go through the event bus as initially implemented
because the event listener in ``AnnotatorComponent.vue`` ( setSampleIds() ) is attached to that component
(because that function needs to know the context of the call - to find which upload to attach the samples names)

* We should not forget to add 'type="button"' attribute to the <button> element in ``SamplerComponent.vue``,
otherwise it will behave like a "Submit" button and would also triggers when the user press enter after entering a new sample ids
in the input field, thus preventing adding further sample ids (because saving causes the panel to close)

* After we save the entered sample ids, the panel should close.
the automated test for the toggling of the panel (controlled by the "samplePanel" variable in ``AnnotatorComponent.vue``)
is not trivial to figure out and requires to handle two cycles of the Vue event loop's Promise-based handler ("Vue.nextTick()")
as it's three steps are needed to get into the context we want to test.
See the 'should close the drawer when clicking save in the sampler' spec in ``AnnotatorComponent.spec.js``