Enable console.log() in Headless Chrome in acceptance tests
When an acceptance test fails, Selenium WebDriver is not showing the output
of the debugging command "console.log()", so we add the necessary Chrome configuration
in the Codeception configuration file for acceptance tests to allow that.

"goog:loggingPrefs" is for version of Chrome 75+, "loggingPrefs" is for version below 75
(like the one we use).