Add missing integration for adding file attributes to upload
Although the meta-data form for upload files (implemented in ``AnnotatorComponent.vue``)
uses the UI component (implemented in ``SpecifierComponent.vue``) for associating
file attributes to the uploaded files
and the PHP controller's action (implemented in ``FilesAnnotateAction.php``) is able to save
attributes posted to it,
there was missing in the form, the input fields that would allow posting of the file attributes.
This is now fixed by adding the necessary hidden input fields in ``AnnotatorComponent.vue``.

Also improved the confirmation message with bootstrap styling and listing of how many attributes have been associated to uploads
(see in ``FilesAnnotateAction.php`` and ``view_profile.php``).

Added an acceptance scenario to end to end test updating upload meta-data and associating file attributes using the form UI.