Fix and simplify SampleIDs UI component for adding samples to upload
In the ``AnnotatorComponent.vue``, Got rid of the "samplesArray" property by doing the deserialization from string to array
in the child component ``SamplerComponent.vue`` and passing the upload's "sample_ids" property directly to that component
instead of that intermediary array property.

This solves a problem whereby the author returns to the meta-data form after it has been completed and saved to the server
with samples attached to the uploads (e.g: after receiving feedback from curators), and wants to add additional samples.

Because the samples UI component is being passed the intermediary array (which starts as empty) and the new samples are assigned
to the upload's string property for samples, the original samples already saved on the server are lost.

By passing instead the upload's string property for samples, the samples UI component is bootstrapped with the samples added the first time.