Create endpoint in FUW API to trigger the job to move files
First we make sure beanstalkd container service exposed port (in ``docker-compose.yml``)
and port configured in FUW API's queue configuration (in ``main.php``) are the same.
We use yii2-queue extension that abstract message queue workflow with its beanstalk driver.

We use the Yii2-flysystem extension, a Yii2 wrapper to the Flysystem  filesystem operations abstraction for the actual file moves job.
Se we need to configure that extension in ``main.php`` config file too. In there, we set up aliases we will be using to refer
to the root directory (@gigadb-data) for the extension
and as base directory for source (@uploads) and destination (@publicftp) during file transfer.

We add Yii2 ActiveRecord relationships to ``FiledropAccount.php`` and ``Upload.php`` as it's needed in the new endpoint controller action
to allow access to uploads from the filedrop account.

Then we configure the new rules for the new endpoint that extends the "filedrop-account" endpoint in ``main.php``
and we create the controller action in ``MoveFilesAction.php``
That's where we create for each upload a worker move job that we push to the message queue.

The job is encapsulated in the class ``MoveJob.php`` that implements Yii2-queue's interface for queued jobs.
In that job class, we use the configured Flysystem component for copying the files,
and for the source and destination we use the configured Yii aliases
(so that changing source and destination is a matter of configuration rather than code change)