Add endpoint in File Upload Wizard to lookup filedrop account records
This is needed by the GigaDB controller for moving files  to public ftp to retrieve the
internal id of filedrop account so it can hit the move files endpoint in File Upload Wizard.

The endpoint will allow lookup by DOI and status so we can get the id of
an active filedrop account associated to that DOI.

We also want to create a method (``getAccounts()``) in ``FiledropService.php`` to connect GigaDB to that new endpoint.
But there was already a method called with name that return all the details of a filedrop account from its internal id.
So we renamed the old method to a clearer ``getAccountDetails()``.