Add GigaDB client to FUW endpoint that triggers files move to public ftp
Functionality implemented in``FiledropService.php``'s ``triggerMoveFiles()``.

In ``DatasetFiledrop.php``, the business object for handling the features that
need to interact with Filedrop accounts, we've added a new ``moveUploadedFiles()``
method that use the aforementioned client method but checks that the dataset
has the correct upload status that allows file to be moved to public ftp.