Correct acceptance tests for move files button
The implementation of the scenario step:
I press "Update Dataset" for dataset "00000*"
wasn't flexible enough. It always assume we want to click the fifth row.
(the row for the first dataset we create in test, usually 000007).
However, to test button state for different upload status (Curation or not)
we need to have two datasets, one for each upload status of interest.
That means that the step need to function properly and actually find the button
on the right row for the right dataset DOI.
That means that the HTML for each row need to contain a DOI that we can query with CSS/XPATH to find a button, typically we use
CSS classes for that hence the addition of the 'rowCssClassExpression' attribute to the CButtonColumn code in ``admin.php``

Then we just have to change the locator on the click() browser automation method to use a valid CSS path using
the two arguments to the step implementation.

Finally, after test run, we need to make sure the new dataset with alternative upload status is removed.
That's done in ``DbExtendedWithHooks.php``

Then we can implement both acceptance scenarios around the (non)existence of the Move Files button.