Add automated tests to file move workers class
it took some times to figure out how to TDD the code that are executed by workers listening to the Yii2-queue because the worker code must implement an interface that get instantiated by the queue listener and the moving of the files is performed
by Flysystem that has several layer of abstractions above the copy instruction.

I found the solution, implemented in MoveJobTest.php,  by doing so:

(1) Make MoveJob.php a Component so we can alter its properties
(2) Create a TestFilesystem.php, that mimics the LocalFilesystem.php class from the Yii2 wrapper of Flysystem and has additional public properties so we can inject test doubles (mocks) so we can expects the MoveJob.php class to call the right method in Flysystem.
(3) Inject a mock of Flystem (with a mocked copy instruction) into the fake TestFilesystem.php and assign that to MoveJob.php’s new property ($_fs)

Also removed the test for existence of source path in MoveJob.php because
Flysystem does that already and raise a FileNotFoundException exception if path
doesn’t exist. (when that happen the message queue detect the exception and send the job to the buried queue)

Also made sure the console app that run the worker have the correct configuration for path aliases and for Flysystem (see in console/config/main.php)