Fix error occuring when pushing to a second queue from a worker
Implement workaround for error:
```
2020-05-28 08:45:09 [-][-][-][error][yii\queue\Queue] [512] backend\models\MoveJob (attempt: 1, PID: 26) is finished with error: yii\base\InvalidConfigException: Queue must be an application component. in /app/vendor/zhuravljov/yii2-queue-monitor/src/JobMonitor.php:227
```

it happened when we have two queues and we use the queue monitoring library (yii2-queue-monitor) and a job is pushed to a queue from within a worker executed by a different queue.

The workaround is to subclass the erring JobMonitor class to override the ways application components are detected, using the tube each queues are attached to for comparisons.

In our use case we are not losing anything in accuracy.