Replace magic numbers by constants from appropriate classes
The FiledropService.php and FileUplodService.php use magic numbers
(raw integers that can't be understood from reading the contextual code)
in order to retrieve the respective File Upload Wizard resources FiledropAccount.php and Upload.php
that have the appropriate status.

So instead we use the meaningful constants from those two model classes.

In order to do that, we first need the GigaDB main app to access the model classes of File Upload Wizard.
We do so by:
* we need to make sure in docker-compose.yml the application and test container services have bind mounted the path of model classes
* adding a PSR-4 block to the composer.json.dist template so that it knows the file path to find the namespaced classes