Add support for sample column in upload of metadata spreadsheet
the bulk upload of file metadata from spreadsheet was ignoring the content of the Sample ID column

The main change happens in DatasetUpload.php to read the "Sample IDs" column in the spreadsheet and merge it
with the upload data's "Sample_ids" attribute for each file of the spreadsheet.

Change were made to sample spreadsheets (and a new one added with examples for Samples) and to the unit test class DatasetUploadTest.php.

I have also added a corresponding acceptance scenario in file-upload-4-author-metadata-bulk.feature.

And fixed an error (dataset not found) in DatasetUpload.php's unit test class caused by fixture data not being loaded.