Add acceptance scenarios for files uploaded using ftp
Although the wizard is the normal for uploading, authors can upload files using an ftp client.
The former approach is covered by acceptance tests scenarios but the latter was missing them.
The scenario starts when the file lands in the incoming directory for ftp upload.

Also doing it as a precursor to refactoring the script that process ftp upload.