Fix scenarios testing samples and attributes buttons on metadata form
The implementation for the scenarios steps clicking the samples or attributes button
on the metadata form were not precise enough causing the button for the wrong file to be clicked
upon the running of acceptance test scenario.

Partial fix was present for the function implementing the clicking on attribute button, but even then
it couldn't distinguish between the sample and button as the CSS selector was limited to the filename as CSS class.
Also the filename as CSS class needed to have the dot escaped otherwise it would denote the extension as another CSS class.

Then we apply the above fixes to the step implementing the clicking on sample button.