Replace ftp upload processing script with more robust console command
Similarly to TusdController.php for file uploaded through Tusd,
Created a FtpController.php Yii2 console command to process ftp file uploads
and create the upload record in FUW database.

Since both commands now works similarly and eventually saved metadata to upload table,
I've ensured the upload factory UploadFactory.php has the factory methods to cater to each scenario.
Hence I've created a createUploadFromFile() method to be used in FtpController()
and replace createUpload() with createUploadFromJSON with similar signature as createUploadFromFile()
to be used in TusdController.php.

Both commands are backed by full automated test suite (unit, functional, acceptance).

I'm now able to remove the flimsy scripts process-uploads.php and file-to-db.php that were created
for the initial prototype.

The ftp console command is executed for each ftp directories that has files
in it in a bash script (process-ftp-uploads.sh) triggered upon filesystem changes.
The BusyBox's internal find command is dumbed down and doesn't have the full range of options of GNU find, so I had
to add GNU find utilities in watcher container service's Dockerfile in order to detect directories with files in it
(that works because the console command moves out the file as part of processing it).

Also moved the acceptance scenario for ftp upload to the feature file
with the corresponding user story (file-upload-2-author-ftp.feature)