Use Flysystem in console command that processes TUS uploads
Flysystem allow us to abstract the filesystem so we can change the type of filestorage
without changing the code. This is important because we are still figuring out which cloud platform
gigadb and FUW will be deployed to.

The subsystem that move files to public ftp already uses Flysystem. Same goes for the processing of ftp file uploads.

The main changes are to replace absolute file path with path relative to the Flysystem's configured root path in code that deal with file operations.

We also needed to move the file copy of the upload file and its associated TUS manifest JSON file from the post-finish shell script
used as hooks to Tusd into TusdController.php so they can make use of Flysys`tem and be covered by the controller functional test suite.