Harmonize docker packaging of worker container services
So that the gigadb-worker container service behaves like the other service of Gigadb main app
and its dockerfile goes alognside the other dockefile for Gigadb services.
fuw-worker container service is a FUW service so it's stay within fuw, however it is moved
from fuw/app/backend to fuw/app/console since is a console based application.