Add fuw related container services to Gitlab deployment stage
in order to deploy FUW on AWS, we are going to reuse the Gitlab config stage
for deploying the prototype (renamed to deploy_fuw for now)

in that stage's config, the main changes are to include support for four new container services:
fuw-public
fuw-worker
gigadb-worker
beanstalkd

and to remove fuw-proto

it was required first to update the docker-compose file that is used to build the production container images (docker-compose.build.yml)

and to make production version of Dockerfile for the workers container service (console/Production-Worker-Dockerfile and packaging/Production-Worker-Dockerfile).

We also need to add the instructions to bring up the container services related to FUW (using the fuw container service)

We also need to update docker-compose file for staging environment to add the four new container services (docker-compose.staging.yml)

Another type of change we need to introduce is provide support for the PSR-4 block of the composer.json files from the various sub-apps.
on Dev, we bind-mount the appropriate source directory but on staging/production source code is baked and not bind-mounted.
So we add COPY instructions in the dockerfiles (the two Production-Worker-Dockerfile, and the main app's Production-Dockerfile) to ensure the model classes mentioned in the PSR-4 block of the composer file can be found.
For the fuw-worker container service, it requires additional gitlab CI steps to have the model class available in its context directory.
We also have the same issue with reference data for file formats and file types that need to be accessible in the two worker container services. The solution is the same as for the PSR-4 block.

Finally we update Nginx config for staging to add a block for proxy passing to the TUSd upload server