Move console gii module config from local config to main config
we want console to have same configuration pattern as backend and frontend
in FUW yii2 advanced template based app.

Also and more importantly, it fixes the yii gii module failed instantiation on staging,
now that gii is configured the same way in all apps.

Also removed the useless YII_ENV_DEV environment in docker-compose.yml as it's a PHP constant.