Fix issue where a development user was used to authenticate REST calls
A mistake in the GigaDB controller action to trigger REST call to FUW for
creating a filedrop account cause the REST authentication to fails
in File Upload Wizard. There are two connected causes:
* the GigaDB action was using hardcoded test user
* the admin user initiating the action didn't have a user identity in File Upload Wizard

I fixed the first item by properly detecting the current logged int user and pass it to the filedrop account
REST client. For the second item, I created a console command for FUW to allow creating new user identities as
an operational action.

I've created a functional test that reproduce the issue. As part of that, I had to had a functional test helper to delete
directories in the filesystem that are created for a filedrop account during the test execution.

To implement the fix it was also necessary to add a role column to the user table in FUW database so that we know which
identity is supposed admin or user.