Workaround multiple listener issue with socat tunnet container
It seems very hard to set up non-SSL an SSL listener for docker daemon
using systemd sockets.
Since directly mounting unix socket in all container is a no-no security wise
we take a middle of the road approach to tunnel the unix socket into a TCP port
using socat running in one container like we do on dev environment.
The difference is the tunnel is started by ansible on production environments.