Remove need for DockerManager init in tusd container service
the tusd console commands processes uploads authors send by web uploading to TUSd.
We need to create an upload object in database keyed against a filedrop account for a given dataset.
FiledropAccount ActiveRecord was used only to find the record and gets its id.
But because it's ActiveRecord it was still initialized (even if it's not instantiated) and therefore an
instance of DockerManager was created requiring a DOCKER_HOST environment variable to be set up.
But because the console command is executed as a hook, accessing environment variable
would be convoluted. So instead we use Yii2 query builder to retrieve the id of the filedrop account.
That approach doesn't require a FiledropAccount object to be initialized.
This fixes the current acceptance failure for the scenario "ftp upload triggers new upload record saved in database"