Fix the Specifier Vue.js component (attributes) life cycle inside Drawer
The Specifier Vue.js component wasn't properly disposed of when the Element Drawer component
embedding it was closed causing inconsistencies in the value of fileAttributes variable
in the UX of that component.
which manifest in attributes appearing for the wrong upload (even though the underlying state
of data was correct).

The solution was to add a ``destroy-on-close`` attribute to the Drawer component so that its child
component are destroy upon the drawer being closed.