Correct workflow when author revisit form after having uploaded files
We should not display the next button in the wizard if file upload aren't complete yet,
but if there have been a previous batch uploaded already, we should display the next button
so that the author can get to the metadata form if they only come to update it.
Also, just "Next" is not helpful enough, we should indicate the next step is for updating metadata.