Replace FUW_FTP_HOST and on use of REMOTE_HOME_URL with REMOTE_HOSTNAME
There are places where we need to refer to the site hostname and not the url.
It was the case for ftp configuration, and now I realise  we need that
for the domain name for generating TLS certificates.
So we deprecate FUW_FTP_HOST and use REMOTE_HOSTNAME wherever we need to refer to the hostname.