Fix acceptance tests scenario for updating gigadb with file infos
the scenario implementation and cleanup didn't prepare the gigadb pre-requisite correctly.
The author can only attach to their files attributes and samples that already exist in GigaDB.
so the test scenarios need to make sure the gigadb database has the right info before the test run
and that the created entries are removed properly afte rthe test run.