Add script to setup test database and remove test binary dump
It will create the database and use CSV-based migrations to populate it
and then create a binary dump from that database.
It allows us to remove the need of a binary dump in the repository for the
test database and unit tests.

In order to not have to alter environment variable for database settings,
we configure a second application component for a CDbConnection class in Yii
(``console.php.dist``).
Then when running the migration, we just add a ``--connectionID`` parameter
keyed on that secondary connection to
the Yii migration command.