Fix error building test container service and running test coverage
When INSTALL_XDEBUG=true is set, it fails with

pecl/xdebug requires PHP (version >= 7.2.0, version <= 8.0.99), installed version is 7.1.33

This is because xdebug 3.0.0 won't work with 7.1 anymore. Solution is to explicitly
install 2.9.8 version

It fixes also the failure generating test coverage that doesn't work when INSTALL_DEBUG
is set to false.