Fix command to manage smoke test data
We need to make sure we use a dataset id very high to avoid collision with existing dataset.

Also when we need the test dataset, we need to remove associated records in file, file_attributes and file_sample table.

Finally, the removal of directories on the file system is deleted as that will be done in
a FUW console command to remove the FUW dropbox