Remove systemd for managing beanstalkd workers
systemd configuration is complex and I was unable to consistently get the
scripts' logs out.
It turns out, it is not necessary to use systemd inside the container
as the command for running the worker script is daemonized by virtue of
running in a detached container and that's good enough for now.