Add JPG file format to prod_like CSV files
The JPG file format was lost when we move to CSV files based test data.
This caused the last part of the file upload workflow to fail on file of that formats.

Fixed by adding the missing data in the corresponding CSV format with the id it had
in the original production_like.pgdmp.