Fix issues where the wrong filedrop account was queried
For audit purpose, filedrop_account record are not deleted, but their status is
changed to 0 (TERMINATED) instead.

When a filedrop_account is created and in used, its status is 1 (ACTIVE)

However there are two components (initial creation of the upload record after files have been uploaded
and housekeeping script to retire a filedrop account) of the file upload workflow that were querying
the database for a filedrop_account record but forgot to select the active status
resulting in sometimes getting and old terminated version of the record.