Improve sequence fixing tool to use DB Schema reflection and be generic
When the database state is created through distinct but combined programmatically
generated schema dump and data dump, the table sequence index can deviate from the highest id used
in the data causing SQL errors at insert time, as the DB generated auto-incremented Ids
that already in-use.
Also the PostgreSQL can occasionally go out-of-sync after restoring binary dump.
To fix both issues, the SequenceFixerCommand was created, initially for the 'curation_log' table.
It works by recalculating the highest used id and assigning that integer+1 as the sequence's ``nextval``
value.

This commit will generalize the fix to all tables, and instead of hard-coding the table names, we use Yii2 Db
connection to give us the list dynamically.