Fix configuration of PHP version to use for application container
There are two compounded issues.

The production Dockerfile was using two different variables for the same purpose.
And the docker compose for building production images is not picking up the php version set as an environment variable.

This problem was expose only now because I think a recent update to composer started to validate php version
and throw error when mismatch.