Fix how the environment for rclone are generated
because at the moment one of the variable is a bucket name whose suffix is app id
we need to have the file generated through using ``generate_config.sh`` and interpolate the tencent app id
instead of hardcoding it.

Replaced the example file by a .dist template.

Updated the README to mention the above as the way to configure rclone in the tool.