Override GitLab variable for COMPSOE_FILE in files-url-updater pipeline
so that the tests in CI can find the docker-compose.yml for the tool.
Without this change it was trying to pick up the value of the same variable define in the main Gitlab config witch points
to GigaDB webapp docker-compose file which can be found in the tool's path.