Disable access to File Upload Wizard workflow from web interface
we create a new list of status specifically for web interface, so that
backend processes and automated tests can still use the full list of statuses.
It's only in the web interface that the workflow is hidden.
Also disable acceptance tests for file upload wizards as they are user centric and go
through web interface.