Use default VPC data block,add tags to created resources, adapt to HK
to avoid having to specify the id of the default VPC, we create a data block
to read the default VPC.

Change the Centos 7 AMI id that was from ap-southeast-1 to its counterpart in the HK region (ap-east-1)

Also HK region doesn't have a t2.micro type of instance, so changed that to t3.micro.

Add "Owner" tag to security group and AWS instance.
At the moment it's hardcoded to my AWS username while I'm still debugging.

Removed the need to pass AWS access key and secret key to the terraform script as Terraform can read them
from the default credentials file.