Adapt Terraform to tightened AWS policies
Some policies are conditioned on the value of tags Owner and Name.
So making sure all created resource have tags with these keys. We do so by setting up
default_tags on the aws provider object for Owner and Environment tag keys (to avoid duplication
as they are the same for all resources) and
making sure we have resource level tag for Name key.

The value for the tag Owner needs to be the username of the role used to run the provisioning job.
For security and reliability's sake we need to programmatically retrieve the IAM user's username.
this is what the shell script ``getIAMUserNameToJSON.sh`` does.
That script is called inside an external data source block whose result is then used to set the value
of the Owner tag at the level of aws default_tags block.

Special attention needs to be made for tagging the root volume as that default tags and resources tag don't trickle
down to volume resource, we therefore need to define a ``volume_tags`` attribute in the aws instance resource block.