Simplify ansible host config, ensure ec2 private IP is sent to ansible
make sure terraform expose data for instantiated ec2 instance's private ip
so that terraform-inventory can expose it to Ansible playbook.
we use Output get the private IP out of the terraform's root module

Need to adjust Ansible playbook and role to use the same variable name that is used in Terraform root module.

From a host config's perspective. It's the same variables that are needed on each environments, so no need to duplicate,
they can be set as global (all:vars) the distinction will be at the Ansisble vault file level (one for each environment)