Add terraform initialisation script,adapt process to live environment
the script ``tf_init.sh`` allow safe way to initialise terraform with remote state on gitlab while the local
mirror is held in an environment-specific directory.

This setup makes it hard to operate on the wrong environment when provisionging or destroying infrastructure.

Update the provisioning of AWS security groups to have their name keyed on the target deployment environment.

Updated volume tags when instantiating ec2 instance to have Environment tag and for the name to be keyed on the environment.

Adjusted Ansible host file so enviroment matches terraform envs directory structure

Add envs/staging and envs/live in repo and gitignore everything in it (but the .gitignore file)

This implements task gigascience/gigadb-website#726