Configure logging for all daemonised container services
To avoid duplication, we use yaml anchors in Compose's Extension Fields to define the logging setting once,
then we use yaml's merge feature to inject that setting to all container services that will be started in daemon mode.

We do that for docker compose on local dev environment and on the one for production environments.
The logging settings on the production environments are more conservative as the resources are more constrained there.