Update bastion playbook
There are new steps added to use the output of the
convert_production_db_to_latest_ver.sh script. A step looks for
the latest backup file in the files-url-updater/sql directory which
is copied to the bastion server. Three shell executions are used
to restore a new gigadb database with the backup file.