Allow user to specify backup file for bastion playbook
Previously, the bastion playbook identified the latest database
backup file created by files-url-updater tool to be copied onto
the bastion server for restoring a Postgresql database on the
RDS. This has now been changed so that a user is required to
specify the backup file when prompted at the
../../../scripts/tf_init.sh step. This information is then captured
in the ansible.properties file for use when running the bastion
playbook.

The documentation in SETUP_CI_CD_PIPELINE.md has been updated to
reflect the above changes.